Fix type inference for NodeList/EdgeList and instances.search
The contained type was being inferred as Any
